NET-355 examples.nntp.NNTPUtils does not compile
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/net/trunk@1071100 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/examples/nntp/NNTPUtils.java
@@ -45,7 +45,7 @@ public class NNTPUtils {
      * @return Article[] An array of Article
      * @throws IOException
      */
-    public  static List<Article> getArticleInfo(NNTPClient client, long lowArticleNumber, long highArticleNumber)
+    public  static List<Article> getArticleInfo(NNTPClient client, int lowArticleNumber, int highArticleNumber)
     throws IOException {
         Reader reader = null;
         List<Article> articles = new ArrayList<Article>();
